+++ libs/base/src/utils/CFileOutputStream.cpp
@@ -131,7 +131,7 @@ size_t  CFileOutputStream::Write(const void *Buffer, size_t Count)
 	Method for moving to a specified position in the streamed resource.
 	 See documentation of CStream::Seek
  ---------------------------------------------------------------*/
-uint64_t CFileOutputStream::Seek(long Offset, CStream::TSeekOrigin Origin)
+uint64_t CFileOutputStream::Seek(uint64_t Offset, CStream::TSeekOrigin Origin)
 {
 	if (!m_of.is_open()) return 0;
 
@@ -159,7 +159,7 @@ uint64_t CFileOutputStream::getTotalBytesCount()
 
 	uint64_t previousPos = getPosition();
 	uint64_t fileSize = Seek(0,sFromEnd);
-	Seek(static_cast<long>(previousPos));
+	Seek(previousPos);
 	return fileSize;
 }
